Desafio Técnico da Semana 1/requirements.txt
requests==2.26.0
python-decouple==3.4
